🤖 displaynames: Implement VariantDisplayName and refactor single module
- Implement VariantDisplayName and VariantDisplayNameOwned in single/variant.rs.

- Refactor single.rs to only re-export Variant, Region, and Script.

- Refactor singular.rs to use unified Medium markers instead of Long markers.
